added static_cast; JIRA: PGAP-9240
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@102917 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ src/algo/align/nw/nw_aligner.cpp
@@ -609,7 +609,7 @@ CNWAligner::TScore CNWAligner::x_Run()
                     SAlignInOut& data = **ii;
 
                     if(static_cast<unsigned int>(data.GetSpace()) >= 10000000 &&
-                       NW_RequestNewThread(m_maxthreads)) {
+                       NW_RequestNewThread(static_cast<unsigned int>(m_maxthreads))) {
                         
                         CNWAlignerThread_Align* thread = 
                             new CNWAlignerThread_Align(this, &data);
